# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  username: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  username: string;
}

export interface UserProfile {
  username: string;
  echoBalance: number;
  bonusFlares: number;
  sonarLevel: number;
}

export interface ItemPurchase {
  itemType: string;
  itemId: string;
}

export interface PurchaseResult {
  echoBalance: number;
  bonusFlares: number;
  sonarLevel: number;
}

export interface EchoClaim {
  packId: string;
}

export interface EchoClaimResult {
  echoBalance: number;
  bonusFlares: number;
  sonarLevel: number;
  granted: number;
}